requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
alembic
bcrypt
pyjwt
python-jose